refactor(build-ui): update structure
