+++ packages/build-ui/src/pages/agits/_id/management/collections/components/collection_name_popup.rs
