+++ packages/build-ui/src/pages/agits/_id/management/collections/components/filter_sidebar.rs
@@ -1,8 +1,6 @@
 use bdk::prelude::*;
 
-use crate::pages::collection::FilterDropdown;
-
-
+use super::FilterDropdown;
 
 #[component]
 pub fn FilterSidebar() -> Element {
