+++ packages/build-ui/src/pages/agits/_id/management/collections/components/mod.rs
@@ -1,19 +1,13 @@
-pub mod page;
-mod i18n;
-
+pub mod collection_name_popup;
 pub mod filter_dropdown;
 pub mod filter_sidebar;
-pub mod models;
 pub mod new_collection_popup;
 pub mod success_popup;
 pub mod transfer_confirmation_popup;
-pub mod collection_name_popup;
 
-pub  use filter_dropdown::FilterDropdown;
+pub use collection_name_popup::CollectionNameModal;
+pub use filter_dropdown::FilterDropdown;
 pub use filter_sidebar::FilterSidebar;
-pub use models::Artwork;
-pub use models::Collection;
 pub use new_collection_popup::NewCollectionModal;
 pub use success_popup::SuccessModal;
 pub use transfer_confirmation_popup::TransferConfirmationModal;
-pub use collection_name_popup::CollectionNameModal;
