+++ packages/build-ui/src/pages/agits/_id/management/collections/components/success_popup.rs
