+++ packages/build-ui/src/pages/agits/_id/management/collections/components/transfer_confirmation_popup.rs
