+++ packages/build-ui/src/pages/agits/_id/management/collections/i18n.rs
@@ -1,0 +1,64 @@
+use bdk::prelude::*;
+
+translate! {
+    CollectionTranslate;
+    collection: {
+        en: "Collection",
+        ko: "컬렉션",
+    },
+
+    collections: {
+        en: "Collections",
+        ko: "컬렉션 목록",
+    },
+
+    floor_price: {
+        en: "Floor price",
+        ko: "최저가",
+    },
+
+    floor_change: {
+        en: "Floor Change",
+        ko: "최저가 변화",
+    },
+
+    volume: {
+        en: "Volume",
+        ko: "거래량",
+    },
+
+    volume_change: {
+        en: "Volume Change",
+        ko: "거래량 변화",
+    },
+
+    owners: {
+        en: "Owners",
+        ko: "소유자",
+    },
+
+    stock: {
+        en: "Stock",
+        ko: "보유 수량",
+    },
+
+    status: {
+        en: "Status",
+        ko: "상태",
+    },
+
+    new_collection: {
+        en: "New Collection",
+        ko: "신규 컬렉션",
+    },
+
+    total_collection: {
+        en: "Total Collections",
+        ko: "전체 컬렉션",
+    },
+
+    search_by_title: {
+        en: "Search by title",
+        ko: "제목으로 검색",
+    },
+}
