+++ packages/build-ui/src/pages/agits/_id/management/collections/mod.rs
@@ -1,5 +1,8 @@
+mod components;
 mod controllers;
 mod i18n;
 mod page;
+
+mod models;
 
 pub use page::CollectionPage;
