+++ packages/build-ui/src/pages/mod.rs
@@ -6,13 +6,10 @@ mod components;
 mod layout;
 mod page;
 
-mod collection;
-
 mod agits;
 
 pub use _route::*;
 pub use layout::HeaderLayout;
 pub use page::RootPage;
-pub use collection::page::CollectionsPage;
 
 pub use agits::*;
